requirements.txt
Flask==2.2.2
Flask-CORS==3.0.10
pycryptodome==3.15.0